+++ src/assets/js/components/WorkoutLogListItem.vue
@@ -1,17 +1,16 @@
 <template>
-  <router-link :to="'/log/' + workoutLog.id">
-    <div class="card">
+  <div class="card">
+    <router-link :to="'/log/' + workoutLog.id">
       <h2>{{ workout.name }}</h2>
-      {{ prettyDate(workoutLog.date) }}
-    </div>
-  </router-link>
+    </router-link>
+    {{ prettyDate(workoutLog.date) }}
+    <button @click="removeWorkoutLogHandler(workoutLog)">x</button>
+  </div>
 </template>
 
 <script>
 import { prettyDate } from "../mixins/prettyDate";
-import { createNamespacedHelpers } from "vuex";
-
-const { mapGetters } = createNamespacedHelpers("workouts");
+import { mapActions, mapGetters } from "vuex";
 
 export default {
   name: "WorkoutLogListItem",
@@ -25,9 +24,22 @@ export default {
     }
   },
   computed: {
-    ...mapGetters(["getWorkoutById"]),
     workout() {
-      return this.getWorkoutById(this.workoutLog.workoutType);
+      return this.$store.state.workouts.workouts[this.workoutLog.workoutType];
+    },
+    ...mapGetters("exerciseLogs", ["getExerciseLogsByWorkoutLogId"]),
+    exerciseLogs() {
+      return this.getExerciseLogsByWorkoutLogId(this.workoutLog.id);
+    }
+  },
+  methods: {
+    ...mapActions("exerciseLogs", ["removeExerciseLog"]),
+    ...mapActions("workoutLogs", ["removeWorkoutLog"]),
+    removeWorkoutLogHandler: function(workoutLog) {
+      this.exerciseLogs.forEach(function(exerciseLog) {
+        this.removeExerciseLog(exerciseLog);
+      }, this);
+      this.removeWorkoutLog(workoutLog);
     }
   }
 };