+++ CHANGELOG.md
@@ -5,12 +5,19 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
 ### Added
 - Version tab that shows current version and changelog
 
+### Changed
+- Workouts are now editable
+- Exercises are now editable
+- Logs are now editable
+
 ## 0.1.0 - 2018-12-06
+
 ### Added
-- Workout Log tab that shows logged workouts and the exercises completed during each workout
+- Log tab that shows logged workouts and the exercises completed during each workout
 - Exercises view tab shows available exercises
 - Workouts tab that shows available workouts and the exercises belonging to them